+++ app.js
@@ -4,20 +4,20 @@ const morgan = require('morgan');
 const tourRouter = require('./routes/tourRoutes');
 const userRouter = require('./routes/userRoutes');
 
-const checkId = (req, res, next, val) => {
-  exports.id = val * 1;
-  exports.tour = tours.find((el) => el.id === id);
+// const checkId = (req, res, next, val) => {
+//   exports.id = val * 1;
+//   exports.tour = tours.find((el) => el.id === id);
 
-  if (!tour) {
-    return res.status(404).json({
-      status: 'fail',
-      message: 'Invalid ID',
-    });
-  }
+//   if (!tour) {
+//     return res.status(404).json({
+//       status: 'fail',
+//       message: 'Invalid ID',
+//     });
+//   }
 
-  req.tour = tour;
-  next();
-};
+//   req.tour = tour;
+//   next();
+// };
 
 const app = express();
 
@@ -31,7 +31,7 @@ app.use(express.json());
 app.use(express.static(`${__dirname}/public`));
 
 // ROUTES
-app.param('id', checkId);
+// app.param('id', checkId);
 app.use('/api/v1/tours', tourRouter);
 app.use('/api/v1/users', userRouter);
