+++ package.json
@@ -28,5 +28,8 @@
     "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-react": "^7.33.2",
     "prettier": "^3.1.1"
+  },
+  "engines": {
+    "node": ">=10.0.0"
   }
 }
